共通のカードレイアウトを作成
