+++ frontend/src/components/card/common/CardContainer.css
@@ -1,0 +1,20 @@
+/* src/components/card/CardContainer.css */
+.card-ui {
+  background: #fafafa;
+  border: 1px solid #e5e5e5;
+  border-radius: 14px;
+  padding: 18px;
+  box-shadow: 0 2px 6px rgba(0, 0, 0, 0.04);
+  max-width: 360px;
+  margin: 0 auto 24px;
+
+  /* 🔽 共通責務：はみ出し完全防止 */
+  word-break: break-word;
+  overflow-wrap: anywhere;
+  white-space: pre-wrap;
+}
+
+/* 操作感を持たせない */
+.card-ui * {
+  cursor: default;
+}
