+++ frontend/src/components/card/common/CardContainer.jsx
@@ -1,0 +1,7 @@
+// src/components/card/CardContainer.jsx
+
+import "./CardContainer.css";
+
+export default function CardContainer({ children }) {
+  return <div className="card-ui">{children}</div>;
+}
