+++ frontend/src/pages/schedules/ScheduleDetailPage.jsx
@@ -9,7 +9,7 @@ import { useDateTime } from "../schedules/useDateTime";
 import { useTodo } from "../todos/useTodo";
 import TodoForm from "../todos/TodoForm";
 
-import ScheduleCard from "../../components/ScheduleCard";
+import ScheduleCard from "../../components/card/schedule/ScheduleCard";
 import ScheduleForm from "./ScheduleForm";
 
 export default function ScheduleDetailPage() {
@@ -30,7 +30,7 @@ export default function ScheduleDetailPage() {
   const [todoForm, setTodoForm] = useState({
     title: "",
     priority: "",
-    due_date: "" || null,
+    due_date: null,
   });
   const [showTodoForm, setShowTodoForm] = useState(false);
 
@@ -73,13 +73,17 @@ export default function ScheduleDetailPage() {
             <div>
               <TodoForm
                 formData={todoForm}
-                onChange={(e) =>
-                  setTodoForm({ ...todoForm, [e.target.name]: e.target.value })
-                }
+                onChange={(e) => {
+                  const { name, value } = e.target;
+                  setTodoForm({
+                    ...todoForm,
+                    [name]: name === "due_date" && value === "" ? null : value,
+                  });
+                }}
                 onSubmit={(e) => {
                   e.preventDefault();
                   handleTodoCreate(todoForm);
-                  setTodoForm({ title: "", priority: "", due_date: "" });
+                  setTodoForm({ title: "", priority: "", due_date: null });
                   setShowTodoForm(false); // 送信後に戻る
                 }}
                 onCancel={() => setShowTodoForm(false)}
